Add chat example.
