+++ app/templates/chat.hbs
@@ -1,0 +1,5 @@
+<form>
+  {{input type="text" value=name placeholder="Name"}}
+  {{input type="text" value=newMessage placeholder="Enter a message"}}
+    <input type="submit" {{action 'chatMessage'}} />
+</form>
