--- app/templates/socket-test.hbs
@@ -1,2 +1,0 @@
-{{name}} {{age}}
-<button name="button" {{action "cherryPickName"}}>Pick Name</button>
