+++ app/views/chat.js
